requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==40.0.1
Deprecated==1.2.13
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
logger==1.4
multidict==6.0.4
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
PyYAML==6.0
requests==2.28.2
smmap==5.0.0
typing-extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
yarl==1.8.2
